Move AirportPlacementArgs to CreateAirport.h
