+++ src/OpenLoco/src/GameCommands/GameCommands.cpp
@@ -1,4 +1,5 @@
 #include "GameCommands.h"
+#include "Airports/CreateAirport.h"
 #include "Audio/Audio.h"
 #include "Buildings/CreateBuilding.h"
 #include "Buildings/RemoveBuilding.h"
