+++ src/OpenLoco/src/Windows/Construction/StationTab.cpp
@@ -1,6 +1,7 @@
 #include "Audio/Audio.h"
 #include "Construction.h"
 #include "Drawing/SoftwareDrawingEngine.h"
+#include "GameCommands/Airports/CreateAirport.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Road/CreateRoadStation.h"
 #include "GameCommands/Road/RemoveRoadStation.h"
